finished new config structure, ported operation atlas to new config structure
