--- mission/AAW_operation_atlas_NATO_AAF_AAS_Infantry.Altis/cfgFOB.hpp
@@ -1,8 +1,0 @@
-class cfgFOB {
-    minDistance = 600;
-    maxEnemyPlace = 5;
-    maxEnemyPlaceRadius = 50;
-    ticketValue = 20;
-    maxEnemyCount = 5;
-    maxEnemyCountRadius = 50;
-};
