--- mission/AAW_operation_atlas_NATO_AAF_AAS_Infantry.Altis/cfgKitGroups.hpp
@@ -1,11 +1,0 @@
-class KitGroups {
-    class Unlimited {
-        requiredGroupMembersPerKit = 1;
-    };
-    class Limited {
-        requiredGroupMembersPerKit = 2;
-    };
-    class Specialized {
-        requiredGroupMembersPerKit = 4;
-    };
-};
