--- mission/AAW_operation_atlas_NATO_AAF_AAS_Infantry.Altis/cfgSquadRallyPoint.hpp
@@ -1,9 +1,0 @@
-class cfgSquadRallyPoint {
-    waitTime = 300;
-    minDistance = 100;
-    spawnCount = 9;
-    nearPlayerToBuild = 2;
-    nearPlayerToBuildRadius = 10;
-    maxEnemyCount = 3;
-    maxEnemyCountRadius = 50;
-};
