+++ mission/AAW_vDev.Stratis/cfgLogisticEast.hpp
@@ -1,4 +1,4 @@
-class cfgLogistic {
+class CSAT {
     objectToSpawn[] = {"mainBaseFlagEast"};
     class MedicalBox {
         displayName = "Medical Box";
