+++ mission/AAW_vDev.Stratis/cfgSides.hpp
@@ -9,9 +9,8 @@ class Sides {
         squadRallyPointObjects[] = {{"Land_TentDome_F", {0,0,0}}};
         FOBBoxObject = "B_CargoNet_01_ammo_F";
         FOBComposition = "FOBNATO";
-        #include "kitsWest.hpp"
-
-        #include "cfgLogisticWest.hpp"
+        kits = "NATO";
+        logistics = "NATO";
     };
 
     class East : West {
@@ -24,9 +23,7 @@ class Sides {
         squadRallyPointObjects[] = {{"Land_TentA_F", {0,0,0}}};
         FOBBoxObject = "O_CargoNet_01_ammo_F";
         FOBComposition = "FOBCSAT";
-        #include "kitsEast.hpp"
-
-        #include "cfgLogisticEast.hpp"
-
+        kits = "CSAT";
+        logistics = "CSAT";
     };
 };
