+++ mission/AAW_vDev.Stratis/description.ext
@@ -75,7 +75,14 @@ class AAW {
 
 	#include "cfgFOB.hpp"
 
+    #include "cfgKitGroups.hpp"
+
     #include "ui\common\aawCore.hpp"
+
+    class cfgLogistics {
+        #include "cfgLogisticWest.hpp"
+        #include "cfgLogisticEast.hpp"
+    };
 
 };
