+++ serverAddons/AAW/addons/AAW_Server/Kit/fn_clientInit.sqf
@@ -13,12 +13,13 @@
     Returns:
     None
 */
+/*
 [QGVAR(KitGroups), missionConfigFile >> QPREFIX >> "KitGroups"] call CFUNC(loadSettings);
 {
     [format [QGVAR(Kit_%1), configName _x], _x >> "Kits"] call CFUNC(loadSettings);
     nil
 } count ("true" configClasses (missionConfigFile >> QPREFIX >> "Sides"));
-
+*/
 // When the player changes he takes his kit with him.
 ["playerChanged", {
     (_this select 0) params ["_newPlayer", "_oldPlayer"];
