+++ serverAddons/AAW/addons/AAW_Server/Kit/fn_getKitDetails.sqf
@@ -16,6 +16,6 @@
 */
 params ["_kitName", "_details"];
 
-private _prefix = format [CFGKITS(%1,%2), playerSide, _kitName];
+private _prefix = format [QUOTE(PREFIX/CfgKits/%1/%2), [format [QUOTE(PREFIX/Sides/%1/kits), (str side group CLib_player)], ""] call CFUNC(getSetting), _kitName];
 
 _details apply {[_prefix + "/" + (_x select 0), _x select 1] call CFUNC(getSetting)};
